DBMS MINI PROJECT
The overall aim of Anish Pharmacy Management System is to automate the existing manual
system by the help of computerized equipment and fully-fledged computer software, 
fulfilling their requirements, so that their valuable data/information can be stored for a longer 
period with easy accessing and manipulation of the same. The required software and 
hardware are easily available and easy to work with.
Anish Pharmacy Management System, as described above, can lead to error free, secure, 
reliable & fast management system. It can assist the user to concentrate on their other 
activities rather to concentrate on the record keeping. Thus, it will help our client in better 
utilization of resources. The client can maintain computerized records without redundant 
entries. That means that one need not be distracted by information that is not relevant, while 
being able to reach the information.
The main objective of this system is it tracks all the information and details of the pharmacy. 
We have developed all type of CRUD (Create, Read, Update and Delete) operations. This is a 
role-based system where admin can perform each and every operation on data