3D MODEL OF DOREMON USING OPEN GL
In this project the main aim is to illustrate the concepts and usage of 
pre-built functions in OpenGL. Creating a cartoon character by the name 
“DOREMON” where the idea behind this project is to display cartoon 
character doremon with computer graphics. This graphics package is based on 
the OpenGL library functions. The programming language used here is C 
using OpenGL libraries. In this project, we are demonstrating a cartoon 
character DOREMON in 3D
To make it look almost exactly the same as the cartoon took as a 
reference, we have used shapes like spheres and cubes and we have used basic 
red, blue, black, white colors 
As it is a 3D display of the cartoon character, we will be able to rotate 
it along the x axis, even we are demonstrating translation, scaling functions, 
zoom in, zoom out function with keyboard interaction using D and K keys